fix(build): reorder CMake statements to enable IPO
