--- src/otserv.h
@@ -1,10 +1,0 @@
-// Copyright 2023 The Forgotten Server Authors. All rights reserved.
-// Use of this source code is governed by the GPL-2.0 License that can be found in the LICENSE file.
-
-#ifndef FS_OTSERV_H
-#define FS_OTSERV_H
-
-void printServerVersion();
-void startServer();
-
-#endif
